requirements.txt
python-telegram-bot[job-queue]==21.6
openpyxl==3.1.5
python-dotenv==1.0.1
supabase==2.5.1
fastapi==0.111.0
uvicorn==0.30.1
requests==2.32.3
